\name{collectIntactPPIData}
\alias{collectIntactPPIData}

\title{A function that parses the Intact repository for bait to prey affiliation}
\description{
 This function is the parsing function for the Intact downloaded data-sets.
}
\usage{
collectIntactPPIData(intactID)
}

\arguments{
  \item{intactID}{A character vector of the Intact ID's that reference
    the particular experiment needed to be parsed.}
}
\details{
  The collectIntactPPIData is a general function that reads the R data file
  tableList.rda and gathers information for the set of
  desired experiments.

  The tableList.rda is generated by running the script parseIntAct.R to
  download and generate a data file will the latest data from IntAct.
}
\value{

  
  \item{allBaits}{A character vector of all distince baits over all experiments}
  \item{allPreys}{A character vector of all distinct preys over all experiments}
  \item{indexSetAll}{A list of lists. Each element of the top list
    corresponds to a particular experiment; the elements of these
    sub-lists are character vectors. The character vectors will always
    carry the bait protein in the first element and all of the preys
    that particular bait has found as the rest of the elements to the
    vector. Note that the multiplicity is not guarenteed to be unity.}
  \item{baitsSystematic}{A character vector with a number of baits
    mapped to their systematic names}
  \item{preysSystematic}{A character vector with a number of preys
    mapped to their systematic names}
  \item{shortLabel}{A character vector listing the first author of each experiment.}
  
}
\references{http://www.ebi.ac.uk/intact}
\author{T Chiang}
\examples{

  inTactData <- collectIntactPPIData(c("EBI-375746", "EBI-531419",
                                    "EBI-295760", "EBI-698096"))

}
\keyword{datagen}

