\name{intActPPIData}
\alias{intActPPIData}
\docType{data}
\title{A data set documenting the output of collectIntactPPIData}

\usage{data(intActData)}
\format{
  The format is:
 chr "y2h"
}
\description{
  This data set is the output for the function collectIntactPPIData. Please
  refer to that man page of that function for a full description of the data set.
}
\examples{
data(intActData)

}
\keyword{datasets}
