\name{sWAC2Sys}
\alias{sWAC2Sys}
\docType{data}
\title{A data set that contains the translation between Swiss accension
  codes to Gene Systematic Names}

\usage{data(sWAC2Sys)}
\format{
  The format is:
 chr "sWAC2Sys"
}
\description{
  This is a named character vector. The names are the Swiss Acension
  Codes and the elements are the corresponding gene systematic names
}
\examples{
data(sWAC2Sys)
}
\keyword{datasets}
